internal: rename album to album_name
